+++ test/test_resources_stores.jl
@@ -33,3 +33,4 @@ run(sim)
 
 @test_throws ErrorException unlock(sto)
 @test_throws ErrorException lock(sto)
+@test_throws ErrorException trylock(sto)